Now LNSSearcher always unassigns the optimization variable.
"optVariable" is always added to the list of variables to unassign. There's an implicit assumption that "optVariable" is the last variable; if this assumption proves to be wrong, some changes might be needed. Since the last variable is always unassigned, the generated random indexes will exclude that value.

Also added a debug message for when an improving solution is found.